requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
WTForms==3.1.2
email-validator==2.1.0.post1
yfinance==0.2.36
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
werkzeug==3.0.1
ollama==0.1.7
requests==2.31.0
